- Fix #489 - librt.so should only be used on Linux
When working with non GA builds of Ingres it's possible that the message 
"undefined symbol: clock_gettime" is seen when building the driver. This 
occurs when  libcompat.1.so is not linked against librt.so. To alleviate
 this problem the ruby driver is linked against librt.so. This however 
causes problems for platforms such as Mac OS X which do not have librt.so


git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@2145 45b5d43f-8932-4c86-835a-3654e5842839