This changes updates support in the Ingres Ruby on Rails adapter for ActiveRecord 2.3.x - Bug #408.
A summary of the changes is as follows:

- Add support for handling the DECIMAL type
- Fix up primary key detection with sequence numbers.
  This assumes that the primary key column is defined using the 
  following syntax:
      col integer not null default sequence.nextval primary key
- Added the configuration option ':use_sequence_for_identity'
- Eliminate ORDER BY in a sub-SELECT for an UPDATE since Ingres does
  not support them in an UPDATE statement
- Ingres does not support "col IN (NULL)" as part of a where clause
  this has been re-written to be "col is NULL"
- Add the database schema for running the ActiveRecord 2.3 unit tests
  against Ingres 9.1/9.2


= Corrections to SQL =

There have been a number of changes to support Ingres's SQL dialect in
particular the handling of NULL values and an ORDER BY clause for an
UPDATE statement ?!??

= Decimal support =

Several changes were required to handle decimal values. Firstly a
mapping was required between the Ruby :decimal type and the Ingres
DECIMAL type. In addition when creating tables Ingres does not support
the following syntax :
  colname DECIMAL DEFAULT xx.yy
where xx.yy is the precision/scale

= Primary key detection = 

The code for detecting the primary key column for a table has been
improved. Previously the adapter would issue 3 SQL queries to
determine the PRIMARY KEY column when a single query against iicolumns
and iiconstraints.

= DEFAULT sequence.nextval =

The adapter has been updated to optionally work with sequence numbers
used with the default value of a column, i.e.:

    id INTEGER NOT NULL DEFAULT sequence_name.NEXTVAL PRIMARY KEY

When the configuration option, in config/database.yml,
':use_sequence_for_identity' is set to TRUE the adapter makes use of
the sequence for the PRIMARY KEY column. If the configuration
parameter is set to false or not defined the adapter selects MAX(id) +
1 as the identity to be used. Since prefetch_primary_key?
returns true we have to manually fetch the next value from the
sequence and then determine if that value is > MAX(id). If not, we keep
fetching the next value for the sequence until we satisfy > MAX(id).
It should be noted there is a potential problem with using
sequence numbers this way. Firstly, since the driver only checks the
value of MAX(id) once it is possible for another session to insert an
identity value beforehand. At this point the application will receive
a "Duplicate key on insert" message from the driver. The alternate
method is to check MAX(id) each time the next sequence number is
fetched. However with this case it's possible that a race condition
will occur as the session is continuously trying to get the next
"free" sequence number that will not cause a duplicate value to be
inserted. 

Support in Ingres for LAST_IDENTITY will alleviate this problem as we
can acquire the actual value inserted rather than calculate which
value might be inserted.

== ActiveRecord test Schema ==

This change provides the ActiveRecord schema used for running the unit
tests, allowing for anyone to be able to run the test suite.


git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@2186 45b5d43f-8932-4c86-835a-3654e5842839