- Update the documentation using RDoc
The standard method for documenting a Ruby project such as the 
Ingres Ruby Driver is to use the RDoc tool. This tool is run 
against the source code for a module/extension, parsing each function
to generate documentation based on the actual source code and it's
comments. This makes maintaining the documentation easier since the 
comments should always reflect the source code.

This change replaces the existing documentation, README and readme.html,
with a single documentation directory created with the following command:

rdoc -U -t "Ingres Ruby Documentation" -o doc --inline-source \
    --format=html -T hanna Ingres.c README.rdoc CHANGELOG  tests/*

    The resulting documentation is placed in the "doc" directory (specified
    using the -o flag).


git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@2258 45b5d43f-8932-4c86-835a-3654e5842839