Fix Bug #555 - SEGV when fetching LOB data
Overview
--------

The Ruby driver for Ingres uses a common function for retrieving all data
from Ingres. In the main this is ok as the majority of types can be fetched
in a single call to IIapi_getColumns(). The only exception to this rule are
the LOB data types, LONG VARCHAR, LONG NVARCHAR and LONG BYTE. With these
types Ingres/OpenAPI streams the data from the server in fixed size chunks.

This change corrects two problems with the code with respect to handling of
LOB data:
- If the size the LOB was less than the segment size then getColumn() would
return the value of dataValue->dv_value as-is without adjusting for the first
two bytes, which contain the length of the LOB/segment
- If the LOB required an additional call to IIapi_getColumns() the driver
would SEGV when using memcpy() to copy the data from dataValue->dv_value
in to buffer. The SEGV was the result of a buffer overflow with
bufferPtr pointing at the end of the buffer and not the end of the last
segment copied.

git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@3370 45b5d43f-8932-4c86-835a-3654e5842839