Fix #651 - Add support for building a Ruby Gem
Overview
--------

Most Ruby extensions are available for download from
http://rubyforge.org and can be downloaded, compiled and installed
using the Ruby Gem (http://rubygems.org/) package manager. For example
to install the mysql driver the following command is used:

gem install mysql

Assuming the necessary header and library files are available the gem
package manager will compile and install the driver in the relevant
location.

In comparison The Ingres Ruby driver and Ruby on Rails (RoR) adapter
requires 8 or steps to download, build and install. This change
updates the Ingres Ruby driver and RoR adapter so that a Ruby gem can
be generated using the following command:

rake gem

In addition zip and tgz packages, suitable for ESD, can be generated
using:

rake package

The resulting archives for the above two commands are placed in a pkg
directory, created in the current working area.

In order to produce a working gem that can be used with RoR, the
directory structure for the source code in SVN will change. The source
code for the driver will move into an extension directory, 'ext', with
the driver tests available as a subdirectory. The RoR related files
move into a directory structure that matches that of the activerecord
gem, i.e.: lib/active_record/connection_adapters

When the gem, activerecord-ingres-adapter, is installed the following
directory structure is created:

activerecord-ingres-adapter-1.4.0
├── ext
│   └── tests
└── lib
    ├── active_record
    │   └── connection_adapters
    ├── arel
    │   └── engines
    │       └── sql
    │           └── compilers
    ├── schema
    └── tests
        └── connections
            └── native_ingres

No additional steps are required to make use of the ActiveRecord
adapter beyond those normally employed when developing a RoR
application. Previously the connection adapter needed to be copied to
the correct location before it could be used. When a reference is made
ingres adapter the Rails app will look for an
activerecord-ingres-adapter directory in the gem path (see the output
from "gem env") and then call lib/activerecord-ingres-adapter.rb. That
file, in turn, includes/requires the ingres_adapter located in
lib/active_record/connection_adapters which provides the RoR
functionality for Ingres.

No changes have been made to the RoR adapter to support Rails 3.0.
That is a separate project and will be dealt with at a later date.

Whilst it's not supported right now, the intention is to update the
gem process to release a new gem to http://rubyforge.org. This would
simplify the install process to a single command (assuming II_SYSTEM
is setup):

gem install activerecord-ingres-adapter

Until such time the generated gem needs to be manually downloaded and
installed, like so:

gem install activerecord-ingres-adapter-1.4.0.gem

A test gem is available for download from
http://dl.dropbox.com/u/2191213/ruby/activerecord-ingres-adapter-1.4.0.gem

git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@3713 45b5d43f-8932-4c86-835a-3654e5842839