Fix #653 - Update s.date dynamically for each release
Overview
--------

This change updates the date specification in Rakefile to update each
time a gem package is generated. Currently the date is a static string
and requires the release maintainer(s) update this string every time a
release is made. Making it dynamic at gem generation time reduces the
burden on the release maintainer(s).

git-svn-id: http://code.ingres.com/ingres/drivers/ruby/main@3722 45b5d43f-8932-4c86-835a-3654e5842839